pyproject.toml
[project]
name = "chat-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cohere>=5.18.0",
    "python-dotenv>=1.0.0",
    "fastapi[standard]>=0.117.1",
]
